+++ java/com/samourai/wallet/api/backend/beans/UnspentOutputComparator.java
@@ -1,11 +1,7 @@
 package com.samourai.wallet.api.backend.beans;
 
-import java.util.Comparator;
+import com.samourai.wallet.utxo.UtxoDetailComparator;
 
-public class UnspentOutputComparator implements Comparator<UnspentOutput> {
-
-  @Override
-  public int compare(UnspentOutput o1, UnspentOutput o2) {
-    return o1.value - o2.value > 0 ? 1 : -1;
-  }
+public class UnspentOutputComparator extends UtxoDetailComparator {
+    // see UtxoDetailComparator
 }