+++ java/com/samourai/wallet/bipWallet/WalletSupplier.java
@@ -1,9 +1,8 @@
 package com.samourai.wallet.bipWallet;
 
-import com.samourai.wallet.api.backend.beans.UnspentOutput;
 import com.samourai.wallet.bipFormat.BipFormat;
+import com.samourai.wallet.bipFormat.BipFormatSupplier;
 import com.samourai.wallet.hd.BIP_WALLET;
-import com.samourai.wallet.hd.BipAddress;
 import com.samourai.whirlpool.client.wallet.beans.WhirlpoolAccount;
 
 import java.util.Collection;
@@ -17,13 +16,11 @@ public interface WalletSupplier {
 
   BipWallet getWallet(BIP_WALLET bip);
 
-  BipWallet getWallet(UnspentOutput unspentOutput);
-
-  BipAddress getAddress(UnspentOutput unspentOutput);
-
   BipWallet getWalletByXPub(String pub);
 
   BipWallet getWalletById(String id);
 
   String[] getXPubs(boolean withIgnoredAccounts, BipFormat... bipFormats);
+
+  BipFormatSupplier getBipFormatSupplier();
 }