+++ java/com/samourai/wallet/utxo/UtxoRef.java
@@ -1,0 +1,6 @@
+package com.samourai.wallet.utxo;
+
+public interface UtxoRef {
+    String getTxHash();
+    int getTxOutputIndex();
+}