+++ src/test/java/com/samourai/wallet/client/KeyBagTest.java
@@ -1,10 +1,10 @@
 package com.samourai.wallet.client;
 
-import com.samourai.wallet.api.backend.beans.UnspentOutput;
 import com.samourai.wallet.bipWallet.BipWallet;
 import com.samourai.wallet.bipWallet.KeyBag;
 import com.samourai.wallet.hd.BIP_WALLET;
 import com.samourai.wallet.test.AbstractTest;
+import com.samourai.wallet.utxo.BipUtxo;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -28,11 +28,11 @@ public class KeyBagTest extends AbstractTest {
   @Test
   public void add() throws Exception {
     BipWallet bipWallet = walletSupplier.getWallet(BIP_WALLET.DEPOSIT_BIP84);
-    UnspentOutput utxo1 = utxoProvider.addUtxo(bipWallet, 1111).toUnspentOutputs().iterator().next();
-    UnspentOutput utxo2 = utxoProvider.addUtxo(bipWallet, 2222).toUnspentOutputs().iterator().next();
-    UnspentOutput utxo3 = utxoProvider.addUtxo(bipWallet, 3333).toUnspentOutputs().iterator().next();
-    byte[] key1 = bipWallet.getAddressAt(utxo1).getHdAddress().getECKey().getPrivKeyBytes();
-    byte[] key2 = bipWallet.getAddressAt(utxo2).getHdAddress().getECKey().getPrivKeyBytes();
+    BipUtxo utxo1 = utxoProvider.addUtxo(bipWallet, 1111).toBipUtxos().iterator().next();
+    BipUtxo utxo2 = utxoProvider.addUtxo(bipWallet, 2222).toBipUtxos().iterator().next();
+    BipUtxo utxo3 = utxoProvider.addUtxo(bipWallet, 3333).toBipUtxos().iterator().next();
+    byte[] key1 = utxo1.getBipAddress(walletSupplier).getHdAddress().getECKey().getPrivKeyBytes();
+    byte[] key2 = utxo2.getBipAddress(walletSupplier).getHdAddress().getECKey().getPrivKeyBytes();
     keyBag.add(utxo1, walletSupplier);
     keyBag.add(utxo2, walletSupplier);
 
@@ -45,11 +45,11 @@ public class KeyBagTest extends AbstractTest {
   @Test
   public void addAll() throws Exception {
     BipWallet bipWallet = walletSupplier.getWallet(BIP_WALLET.DEPOSIT_BIP84);
-    UnspentOutput utxo1 = utxoProvider.addUtxo(bipWallet, 1111).toUnspentOutputs().iterator().next();
-    UnspentOutput utxo2 = utxoProvider.addUtxo(bipWallet, 2222).toUnspentOutputs().iterator().next();
-    UnspentOutput utxo3 = utxoProvider.addUtxo(bipWallet, 3333).toUnspentOutputs().iterator().next();
-    byte[] key1 = bipWallet.getAddressAt(utxo1).getHdAddress().getECKey().getPrivKeyBytes();
-    byte[] key2 = bipWallet.getAddressAt(utxo2).getHdAddress().getECKey().getPrivKeyBytes();
+    BipUtxo utxo1 = utxoProvider.addUtxo(bipWallet, 1111).toBipUtxos().iterator().next();
+    BipUtxo utxo2 = utxoProvider.addUtxo(bipWallet, 2222).toBipUtxos().iterator().next();
+    BipUtxo utxo3 = utxoProvider.addUtxo(bipWallet, 3333).toBipUtxos().iterator().next();
+    byte[] key1 = utxo1.getBipAddress(walletSupplier).getHdAddress().getECKey().getPrivKeyBytes();
+    byte[] key2 = utxo2.getBipAddress(walletSupplier).getHdAddress().getECKey().getPrivKeyBytes();
     keyBag.addAll(Arrays.asList(utxo1, utxo2), walletSupplier);
 
     // verify